+++ context_auditor/__init__.py
@@ -8,7 +8,14 @@ from .io import load_jsonl, write_jsonl
 from .mitigation import MitigationReport, remove_exact_duplicate_segments
 from .mitigation_eval import evaluate_mitigation_strategies, final_invocations, summarize_mitigation_rows
 from .provenance import SourceType, label_message_source, segment_messages
-from .providers import ChatProvider, MockChatProvider, OpenAICompatibleChatProvider, provider_from_environment
+from .providers import (
+    ChatProvider,
+    DeepSeekChatProvider,
+    MockChatProvider,
+    OpenAICompatibleChatProvider,
+    provider_environment_status,
+    provider_from_environment,
+)
 from .schema import AuditTrace, Message, Segment
 from .tracer import RuntimeAuditor, TraceMetadata
 
@@ -17,6 +24,7 @@ __all__ = [
     "BloatThresholds",
     "ChatProvider",
     "ContextGrowthGuard",
+    "DeepSeekChatProvider",
     "GuardConfig",
     "Message",
     "MitigationReport",
@@ -33,6 +41,7 @@ __all__ = [
     "load_jsonl",
     "evaluate_mitigation_strategies",
     "final_invocations",
+    "provider_environment_status",
     "provider_from_environment",
     "remove_exact_duplicate_segments",
     "segment_messages",